- Finished Blob Storage - Added Review retrieval in back end

+++ Core/Entities/Review.cs
@@ -5,7 +5,7 @@ namespace Core.Entities
     public class Review : BaseEntity
     {
         public int TicketId { get; set; }
-        public byte Stars { get; set; }
+        public double Stars { get; set; }
         public string? Details { get; set; }
         public Ticket Ticket { get; set; }
     }
